Updated WebApp script for pulling game content now that it is decoupled and located in Microservice for Tetris.
Removed tetris.js code from WebApp and adding the missing code to the Tetris repo's tetris.js file.

Updated TetrisController.cs to use the absolute url that the microservice is hosting it at instead of the relative path that the WebApp was using previously. Need to check if this needs to change to be a dynamic appsettings variable instead.
