+++ Bucstop WebApp/BucStop/wwwroot/playcount.json
@@ -1,1 +1,1 @@
-[{"Id":1,"Info":null,"Title":"Snake","Content":"/js/snake.js","Author":null,"DateAdded":null,"Description":"Snake Description","HowTo":null,"Thumbnail":"/images/snake.jpg","LeaderBoard":null,"PlayCount":0},{"Id":2,"Info":null,"Title":"Tetris","Content":"/js/tetris.js","Author":null,"DateAdded":null,"Description":"Tetris Description","HowTo":null,"Thumbnail":"/images/tetris.jpg","LeaderBoard":null,"PlayCount":0},{"Id":3,"Info":null,"Title":"Pong","Content":"/js/pong.js","Author":null,"DateAdded":null,"Description":"Pong Description","HowTo":null,"Thumbnail":"/images/pong.jpg","LeaderBoard":null,"PlayCount":0}]
+[{"Id":1,"Info":null,"Title":"Snake","Content":"/js/snake.js","Author":null,"DateAdded":null,"Description":"Snake Description","HowTo":null,"Thumbnail":"/images/snake.jpg","LeaderBoard":null,"PlayCount":0},{"Id":2,"Info":null,"Title":"Tetris","Content":"/js/tetris.js","Author":null,"DateAdded":null,"Description":"Tetris Description","HowTo":null,"Thumbnail":"/images/tetris.jpg","LeaderBoard":null,"PlayCount":1},{"Id":3,"Info":null,"Title":"Pong","Content":"/js/pong.js","Author":null,"DateAdded":null,"Description":"Pong Description","HowTo":null,"Thumbnail":"/images/pong.jpg","LeaderBoard":null,"PlayCount":0}]
