+++ Team-3-BucStop_Tetris/Tetris/Controllers/TetrisController.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.Logging;
 
 using Microsoft.Extensions.Caching.Distributed;
 using System.Text.Json;
+using static System.Net.WebRequestMethods;
 
 namespace Tetris
 {
@@ -21,7 +22,10 @@ namespace Tetris
                 Id = 2,
                 Title = "Tetris",
                 Author = "Fall 2023 Semester",
-                Content = "~/js/tetris.js",
+                //Content = "~/js/tetris.js", //changing URL now that the MS hosts game code
+                // need to probably change this url so that it is aware of the environment dynamically?
+                // i think this would mean using one of the app settings variables instead of hard coding
+                Content = "https://localhost:7182/js/tetris.js",
                 DateAdded = "",
                 Description = "Tetris is a classic arcade puzzle game where the player has to arrange falling blocks, also known as Tetronimos, of different shapes and colors to form complete rows on the bottom of the screen. The game gets faster and harder as the player progresses, and ends when the Tetronimos reach the top of the screen.",
                 HowTo = "Control with arrow keys: Up arrow to spin, down to speed up fall, space to insta-drop.",
