+++ Team-3-BucStop_Tetris/Tetris/wwwroot/js/tetris.js
@@ -140,6 +140,11 @@ function placeTetromino() {
             break;
     }
 
+    // Show score to website, updated after the above code runs
+    const playerScore = document.getElementById("playerScore");
+
+    playerScore.textContent = `Current Score: ${score}`;
+
     tetromino = getNextTetromino();
 }
 
@@ -331,5 +336,18 @@ document.addEventListener('keydown', function (e) {
     }
 });
 
-// start the game
-rAF = requestAnimationFrame(loop);
+// will display an instruction for the player in order to trigger the loop of the game
+function showStartScreen() {
+    context.font = '36px Arial';
+    context.textAlign = 'center';
+    context.fillText('Press space to start', canvas.width / 2, canvas.height / 2);
+}
+
+// on keyboard press of space, start the game.
+document.body.onkeyup = function (e) {
+    if (e.keyCode == 32) {
+        rAF = requestAnimationFrame(loop);
+    }
+}
+
+showStartScreen();
